refactor(base-driver): split protocol.ts into focused modules
protocol.ts mixed payload validation, session/protocol resolution, WD-proxy
handling, and Express route wiring in one 590-line file. Split it into
params.ts, session.ts, proxy.ts, and command-names.ts, and renamed the
remainder to route-handler.ts. Decomposed buildHandler's request handler
into named steps (logDeprecationWarning, ensureSessionExists, preparePayload,
buildCommandArgs, runDriverCommand, buildErrorResponse, sendHandlerResponse)
instead of one long closure. Also flags the setNetworkConnection validator
as deprecated, matching its route.

No public API changes; protocol/index.ts re-exports are unchanged.
