+++ packages/base-driver/lib/protocol/command-names.ts
@@ -1,0 +1,5 @@
+export const CREATE_SESSION_COMMAND = 'createSession';
+export const DELETE_SESSION_COMMAND = 'deleteSession';
+export const GET_STATUS_COMMAND = 'getStatus';
+export const LIST_DRIVER_COMMANDS_COMMAND = 'listCommands';
+export const LIST_DRIVER_EXTENSIONS_COMMAND = 'listExtensions';
