+++ packages/base-driver/lib/protocol/validators.ts
@@ -8,13 +8,10 @@ export const validators = {
       throw new Error(`'${util.truncateString(String(url), {length: MAX_URL_ERROR_LENGTH})}' must be a valid URL`);
     }
   },
+  /** @deprecated Only used by the deprecated `setNetworkConnection` MJSONWP route. */
   setNetworkConnection: (type: any) => {
-    if (!isNumber(type) || [0, 1, 2, 4, 6].indexOf(type) === -1) {
+    if (![0, 1, 2, 4, 6].includes(Number(type))) {
       throw new Error('Network type must be one of 0, 1, 2, 4, 6');
     }
   },
 };
-
-function isNumber(o: any): o is number {
-  return typeof o === 'number' || !Number.isNaN(parseInt(o, 10)) || !Number.isNaN(parseFloat(o));
-}
